remove util and some commented code
